Move tests to parse_tests
This avoids having to parse the tests during a
non test build and keeps parsers.rs under the line
length limit for import into mozilla-central.
