+++ glsl/Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "glsl"
-version = "4.0.1"
+version = "4.0.2"
 license = "BSD-3-Clause"
 authors = ["Dimitri Sabadie <dimitri.sabadie@gmail.com>"]
 description = "A GLSL450/GLSL460 parser."
